requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
passlib==1.7.4
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a19
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0.post1
uvloop==0.16.0
watchgod==0.7
websockets==10.1
